i32->i64; day 7

+++ rust/src/years/mod.rs
@@ -2,7 +2,7 @@ mod y2023;
 mod y2024;
 
 
-pub fn run_day(day: &mut u8, year: u16, test: bool) -> (i32, i32) {
+pub fn run_day(day: &mut u8, year: u16, test: bool) -> (i64, i64) {
     match year {
         2023 => {
             y2023::run_day(day, test)
